creating abstract for list container and removing specific id for header bar
