+++ packages/components/e2e/component-objects/src/main/java/org/talend/component/HeaderBar.java
@@ -14,15 +14,7 @@ public class HeaderBar extends Component {
 
     static final String SELECTOR = ".tc-header-bar";
 
-    static final String HELP_SELECTOR = "header-help";
-
-    static final String USER_SELECTOR = "header-user";
-
     static final String USER_DROP_DOWN_SELECTOR = ".dropdown-menu.dropdown-menu-right";
-
-    static final String ABOUT_SELECTOR = "//li/a[@label='About']";
-
-    static final String LOGOUT_SELECTOR = "//li/a[@label='Logout']";
 
     /**
      * AppHeaderBar constructor
@@ -33,23 +25,11 @@ public class HeaderBar extends Component {
         super(driver, NAME, SELECTOR);
     }
 
-    public WebElement getHelp() {
-        return this.getElement().findElement(By.id(HELP_SELECTOR));
-    }
-
-    public WebElement getUser() {
-        return this.getElement().findElement(By.id(USER_SELECTOR));
-    }
-
     public WebElement getUserMenu() {
         return this.getElement().findElement(By.cssSelector(USER_DROP_DOWN_SELECTOR));
     }
 
-    public WebElement getInfo() {
-        return this.getElement().findElement(By.xpath(ABOUT_SELECTOR));
-    }
-
-    public WebElement getLogout() {
-        return this.getElement().findElement(By.xpath(LOGOUT_SELECTOR));
+    public WebElement getItemById(String id) {
+        return this.getElement().findElement(By.id(id));
     }
 }
